Feat: Implement GLTF texture reading
