+++ crates/bevy_gltf/Cargo.toml
@@ -13,10 +13,13 @@ keywords = ["bevy"]
 # bevy
 bevy_app = { path = "../bevy_app", version = "0.1" }
 bevy_asset = { path = "../bevy_asset", version = "0.1" }
+bevy_math = { path = "../bevy_math", version = "0.1" }
 bevy_render = { path = "../bevy_render", version = "0.1" }
+bevy_pbr = { path = "../bevy_pbr", version = "0.1" }
 
 # other
 gltf = { version = "0.15.2", default-features = false, features = ["utils"] }
 thiserror = "1.0"
 anyhow = "1.0"
 base64 = "0.12.3"
+image = "0.23.8"
