+++ crates/bevy_gltf/src/lib.rs
@@ -1,4 +1,6 @@
 mod loader;
+mod gltf_reader;
+
 pub use loader::*;
 
 use bevy_app::prelude::*;
