+++ examples/3d/load_model.rs
@@ -13,6 +13,10 @@ fn setup(
     asset_server: Res<AssetServer>,
     mut materials: ResMut<Assets<StandardMaterial>>,
 ) {
+    let cube_texture = asset_server
+        .load::<Texture, _>("assets/models/cube/textured_cube.gltf")
+        .unwrap();
+
     // add entities to the world
     commands
         // mesh
@@ -34,6 +38,20 @@ fn setup(
                 .unwrap(),
             // create a material for the mesh
             material: materials.add(Color::rgb(0.5, 0.4, 0.3).into()),
+            translation: Translation::new(1.5, 0.0, 0.0),
+            ..Default::default()
+        })
+        // textured mesh
+        .spawn(PbrComponents {
+            mesh: asset_server
+                .load::<Mesh, _>("assets/models/cube/textured_cube.gltf")
+                .unwrap(),
+            // create a material for the mesh
+            material: materials.add(
+                StandardMaterial {
+                    albedo_texture: Some(cube_texture),
+                    ..Default::default()
+                }),
             translation: Translation::new(1.5, 0.0, 0.0),
             ..Default::default()
         })
